requirements.txt
fastapi
uvicorn
pymongo
langchain
openai
python-dotenv